e2e: centralize utilities
move utility function in the main suite file,
to foster code reuse.

Trivial code movement, no expected changes in behavior.

(cherry picked from commit 1a8759d098c1efedc771bb45d09275b70522b32b)
